refactor: extract IMountedEquipmentInstance to @/types/equipment/
