+++ src/components/customizer/armor/shared/ArmorFills.tsx
@@ -6,15 +6,9 @@
 
 import React from 'react';
 
-/**
- * Status thresholds for armor coloring
- */
-export const ARMOR_STATUS = {
-  HEALTHY: { min: 0.6, color: '#22c55e' }, // green-500  (≥60%)
-  MODERATE: { min: 0.4, color: '#f59e0b' }, // amber-500  (≥40%)
-  LOW: { min: 0.2, color: '#f97316' }, // orange-500 (≥20%)
-  CRITICAL: { min: 0, color: '#ef4444' }, // red-500    (<20%)
-} as const;
+import { ARMOR_STATUS } from '@/constants/armorStatus';
+
+export { ARMOR_STATUS };
 
 export const SELECTED_COLOR = '#3b82f6'; // blue-500
 export const SELECTED_STROKE = '#60a5fa'; // blue-400
