+++ src/stores/unitState.ts
@@ -7,6 +7,7 @@
  * @spec openspec/specs/unit-store-architecture/spec.md
  */
 
+import { IArmorAllocation } from '@/types/construction/ArmorAllocation';
 import { ArmorTypeEnum } from '@/types/construction/ArmorType';
 import { CockpitType } from '@/types/construction/CockpitType';
 import { MechLocation } from '@/types/construction/CriticalSlotAllocation';
@@ -27,6 +28,7 @@ import {
 import { RulesLevel } from '@/types/enums/RulesLevel';
 import { TechBase } from '@/types/enums/TechBase';
 import { EquipmentCategory, IEquipmentItem } from '@/types/equipment';
+import { IMountedEquipmentInstance } from '@/types/equipment/MountedEquipment';
 import { MechConfiguration, UnitType } from '@/types/unit/BattleMechInterfaces';
 import {
   getMaxTotalArmor,
@@ -39,89 +41,13 @@ import { JumpJetType } from '@/utils/construction/movementCalculations';
 import { generateUnitId as generateUUID } from '@/utils/uuid';
 
 // =============================================================================
-// Armor Allocation Types
+// Armor Allocation Types (Re-export Facade)
 // =============================================================================
 
-/**
- * Per-location armor allocation
- * Stores armor points assigned to each location (front and rear for torsos)
- * Supports all mech configurations: Biped, Quad, Tripod, LAM, QuadVee
- */
-export interface IArmorAllocation {
-  /** Index signature allowing any MechLocation key */
-  [key: string]: number;
-
-  // === Universal locations (all configurations) ===
-  /** Head armor points */
-  [MechLocation.HEAD]: number;
-  /** Center Torso front armor points */
-  [MechLocation.CENTER_TORSO]: number;
-  /** Center Torso rear armor points */
-  centerTorsoRear: number;
-  /** Left Torso front armor points */
-  [MechLocation.LEFT_TORSO]: number;
-  /** Left Torso rear armor points */
-  leftTorsoRear: number;
-  /** Right Torso front armor points */
-  [MechLocation.RIGHT_TORSO]: number;
-  /** Right Torso rear armor points */
-  rightTorsoRear: number;
-
-  // === Biped/Tripod/LAM arm locations ===
-  /** Left Arm armor points */
-  [MechLocation.LEFT_ARM]: number;
-  /** Right Arm armor points */
-  [MechLocation.RIGHT_ARM]: number;
-
-  // === Biped/Tripod/LAM leg locations ===
-  /** Left Leg armor points */
-  [MechLocation.LEFT_LEG]: number;
-  /** Right Leg armor points */
-  [MechLocation.RIGHT_LEG]: number;
-
-  // === Tripod-specific location ===
-  /** Center Leg armor points (Tripod only) */
-  [MechLocation.CENTER_LEG]: number;
-
-  // === Quad/QuadVee-specific locations ===
-  /** Front Left Leg armor points (Quad/QuadVee only) */
-  [MechLocation.FRONT_LEFT_LEG]: number;
-  /** Front Right Leg armor points (Quad/QuadVee only) */
-  [MechLocation.FRONT_RIGHT_LEG]: number;
-  /** Rear Left Leg armor points (Quad/QuadVee only) */
-  [MechLocation.REAR_LEFT_LEG]: number;
-  /** Rear Right Leg armor points (Quad/QuadVee only) */
-  [MechLocation.REAR_RIGHT_LEG]: number;
-}
-
-/**
- * Create empty armor allocation (all zeros)
- * Includes all possible locations for all mech configurations
- */
-export function createEmptyArmorAllocation(): IArmorAllocation {
-  return {
-    // Universal locations
-    [MechLocation.HEAD]: 0,
-    [MechLocation.CENTER_TORSO]: 0,
-    centerTorsoRear: 0,
-    [MechLocation.LEFT_TORSO]: 0,
-    leftTorsoRear: 0,
-    [MechLocation.RIGHT_TORSO]: 0,
-    rightTorsoRear: 0,
-    // Biped/Tripod/LAM locations
-    [MechLocation.LEFT_ARM]: 0,
-    [MechLocation.RIGHT_ARM]: 0,
-    [MechLocation.LEFT_LEG]: 0,
-    [MechLocation.RIGHT_LEG]: 0,
-    // Tripod-specific
-    [MechLocation.CENTER_LEG]: 0,
-    // Quad/QuadVee-specific
-    [MechLocation.FRONT_LEFT_LEG]: 0,
-    [MechLocation.FRONT_RIGHT_LEG]: 0,
-    [MechLocation.REAR_LEFT_LEG]: 0,
-    [MechLocation.REAR_RIGHT_LEG]: 0,
-  };
-}
+export {
+  type IArmorAllocation,
+  createEmptyArmorAllocation,
+} from '@/types/construction/ArmorAllocation';
 
 /**
  * Convert ArmorAllocationResult from calculation utility to IArmorAllocation format
@@ -244,133 +170,16 @@ export function createEmptySelectionMemory(): ISelectionMemory {
 }
 
 // =============================================================================
-// Mounted Equipment Types
+// Mounted Equipment Types (Re-export Facade)
 // =============================================================================
 
-/**
- * Mounted equipment instance on a unit
- *
- * Each instance represents a single piece of equipment added to the unit.
- * Multiple instances of the same equipment type can exist (e.g., 4 Medium Lasers).
- */
-export interface IMountedEquipmentInstance {
-  /** Unique instance identifier (generated when added) */
-  readonly instanceId: string;
-  /** Reference to equipment definition ID */
-  readonly equipmentId: string;
-  /** Display name of the equipment */
-  readonly name: string;
-  /** Equipment category for grouping */
-  readonly category: EquipmentCategory;
-  /** Weight in tons */
-  readonly weight: number;
-  /** Number of critical slots required */
-  readonly criticalSlots: number;
-  /** Heat generated (for weapons) */
-  readonly heat: number;
-  /** Tech base of the equipment */
-  readonly techBase: TechBase;
-  /** Location where equipment is placed (undefined = unallocated) */
-  readonly location?: MechLocation;
-  /** Critical slot indices in the location (undefined = unallocated) */
-  readonly slots?: readonly number[];
-  /** Whether mounted facing rear arc */
-  readonly isRearMounted: boolean;
-  /** Linked ammunition instance ID (for weapons that use ammo) */
-  readonly linkedAmmoId?: string;
-  /**
-   * Whether this equipment can be removed via the loadout tray.
-   * Configuration components (Endo Steel, Ferro-Fibrous, Jump Jets, etc.)
-   * are managed via their respective tabs and cannot be removed directly.
-   * Defaults to true for user-added equipment.
-   */
-  readonly isRemovable: boolean;
-  /**
-   * Whether this equipment is pod-mounted on an OmniMech.
-   * Pod-mounted equipment can be swapped between configurations.
-   * Fixed equipment (isOmniPodMounted: false) is part of the base chassis.
-   * Only relevant when the unit's isOmni flag is true.
-   */
-  readonly isOmniPodMounted: boolean;
-  /**
-   * Actual crit entries consumed on a superheavy mech.
-   * Equal to criticalSlots for standard mechs, ceil(criticalSlots / 2) for superheavy.
-   * Only set when the unit is a superheavy mech.
-   *
-   * @spec openspec/specs/superheavy-mech-system/spec.md
-   */
-  readonly critEntries?: number;
-}
-
-/**
- * Create a mounted equipment instance from an equipment item
- * @param item The equipment definition
- * @param instanceId Unique instance identifier
- * @param isRemovable Whether the equipment can be removed via the loadout tray (default: true)
- * @param isOmniPodMounted Whether the equipment is pod-mounted on an OmniMech (default: false)
- */
-export function createMountedEquipment(
-  item: IEquipmentItem,
-  instanceId: string,
-  isRemovable: boolean = true,
-  isOmniPodMounted: boolean = false,
-): IMountedEquipmentInstance {
-  return {
-    instanceId,
-    equipmentId: item.id,
-    name: item.name,
-    category: item.category,
-    weight: item.weight,
-    criticalSlots: item.criticalSlots,
-    heat: 'heat' in item ? (item as { heat: number }).heat : 0,
-    techBase: item.techBase,
-    location: undefined,
-    slots: undefined,
-    isRearMounted: false,
-    linkedAmmoId: undefined,
-    isRemovable,
-    isOmniPodMounted,
-  };
-}
-
-/**
- * Calculate total equipment weight
- */
-export function getTotalEquipmentWeight(
-  equipment: readonly IMountedEquipmentInstance[],
-): number {
-  return equipment.reduce((total, item) => total + item.weight, 0);
-}
-
-/**
- * Calculate total equipment critical slots
- */
-export function getTotalEquipmentSlots(
-  equipment: readonly IMountedEquipmentInstance[],
-): number {
-  return equipment.reduce((total, item) => total + item.criticalSlots, 0);
-}
-
-/**
- * Get equipment count by category
- */
-export function getEquipmentByCategory(
-  equipment: readonly IMountedEquipmentInstance[],
-): Record<EquipmentCategory, IMountedEquipmentInstance[]> {
-  const result = {} as Record<EquipmentCategory, IMountedEquipmentInstance[]>;
-
-  // Initialize all categories with empty arrays
-  for (const category of Object.values(EquipmentCategory)) {
-    result[category] = [];
-  }
-
-  // Group equipment by category
-  for (const item of equipment) {
-    result[item.category].push(item);
-  }
-
-  return result;
-}
+export type { IMountedEquipmentInstance } from '@/types/equipment/MountedEquipment';
+export {
+  createMountedEquipment,
+  getTotalEquipmentWeight,
+  getTotalEquipmentSlots,
+  getEquipmentByCategory,
+} from '@/types/equipment/MountedEquipment';
 
 // =============================================================================
 // Unit State Interface
